requirements.txt
Django~=1.11
djangorestframework==3.8.2
coreapi==2.3.3
-e .
